Relocate session restoration to Root
